requirements.txt
# Core ML dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0

# Configuration and data handling
PyYAML>=6.0
nibabel>=4.0.2  # Medical imaging (NIfTI format)

# Image processing and augmentation
scikit-image==0.20.0
scipy>=1.9.0

# Visualization dependencies
tqdm==4.64.0
matplotlib==3.7.1
seaborn==0.12.2

# Optional: For mobile deployment (commented out for basic training)
# onnx>=1.12.0
# onnxruntime>=1.12.0
# tensorflow-lite>=2.10.0
